+++ src/commands/mc.workspace.refresh.ts
@@ -1,0 +1,5 @@
+import { CurrentPackProvider } from "../class/CurrentPackProvider";
+
+export default (provider: CurrentPackProvider) => {
+    provider.refresh();
+};